+++ RELEASENOTES.md
@@ -14,6 +14,9 @@ Release notes
         ([#10604](https://github.com/google/ExoPlayer/issues/10604)).
     *   Add `ExoPlayer.Builder.setPlaybackLooper` that sets a pre-existing
         playback thread for a new ExoPlayer instance.
+*   Session:
+    *   Add helper method to convert platform session token to Media3
+        `SessionToken` ([#171](https://github.com/androidx/media/issues/171)).
 *   Remove deprecated symbols:
     *   Remove `DefaultAudioSink` constructors, use `DefaultAudioSink.Builder`
         instead.